+++ samples/run.exs
@@ -1,3 +1,5 @@
+# It is possible to use multiple formatters so that you have both the Console
+# output and a csv file.
 list = Enum.to_list(1..10_000)
 map_fun = fn(i) -> [i, i * i] end
 
@@ -5,5 +7,8 @@ Benchee.run(%{
   "flat_map"    => fn -> Enum.flat_map(list, map_fun) end,
   "map.flatten" => fn -> list |> Enum.map(map_fun) |> List.flatten end
 },
-  formatters: [Benchee.Formatters.CSV],
+  formatters: [
+    Benchee.Formatters.CSV,
+    Benchee.Formatters.Console
+  ],
   formatter_options: [csv: [file: "my.csv"]])